Moved serving logic into src folder for frontend package
