+++ packages/frontend/package.json
@@ -15,6 +15,7 @@
     "@openapi-platform/config": "1.0.0-alpha.0",
     "@openapi-platform/frontend-dist": "1.0.0-alpha.0",
     "@openapi-platform/logger": "1.0.0-alpha.0",
+    "connect-history-api-fallback": "^1.5.0",
     "source-map-support": "^0.5.9",
     "webpack": "^4.17.1"
   }
